Split primsa client up
